+++ click_extended/__init__.py
@@ -2,8 +2,10 @@
 
 from click_extended.core._child_node import ChildNode, ProcessContext
 from click_extended.core._global_node import GlobalNode
+from click_extended.core._node import Node
 from click_extended.core._parent_node import ParentNode
 from click_extended.core._root_node import RootNode
+from click_extended.core._tree import Tree
 from click_extended.core.argument import argument
 from click_extended.core.command import command
 from click_extended.core.env import env
@@ -15,8 +17,10 @@ __all__ = [
     "ChildNode",
     "ProcessContext",
     "GlobalNode",
+    "Node",
     "ParentNode",
     "RootNode",
+    "Tree",
     "argument",
     "command",
     "env",