+++ click_extended/types.py
@@ -1,10 +1,9 @@
 """Types used in `click_extended` which can be useful for users."""
 
-from click_extended.core._child_node import ProcessContext
-from click_extended.core._node import Node
-from click_extended.core._parent_node import ParentNode
-from click_extended.core._root_node import RootNode
-from click_extended.core._tree import Tree
+# pylint: disable=invalid-name
+
+from typing import TYPE_CHECKING, Any, Callable
+
 from click_extended.core.argument import Argument
 from click_extended.core.command import Command
 from click_extended.core.env import Env
@@ -12,17 +11,17 @@ from click_extended.core.group import Group
 from click_extended.core.option import Option
 from click_extended.core.tag import Tag
 
+if TYPE_CHECKING:
+    from click_extended.core._parent_node import ParentNode
+
 Tags = dict[str, Tag]
 Siblings = list[str]
-Parent = ParentNode | Tag
+Parent = "ParentNode | Tag"
+Decorator = Callable[[Callable[..., Any]], Callable[..., Any]]
 
 __all__ = [
-    "Node",
+    "Decorator",
     "Parent",
-    "ParentNode",
-    "ProcessContext",
-    "RootNode",
-    "Tree",
     "Argument",
     "Command",
     "Env",